Added UserProfile and replaced some funcs also changed db settings

+++ src/main/java/com/example/computerStock/domain/Message.java
@@ -7,12 +7,9 @@ import javax.persistence.criteria.Fetch;
 @Entity
 public class Message {
 
-
-
-
     @Id
     @GeneratedValue(strategy=GenerationType.AUTO)
-    private Integer id;
+    private Long id;
     private String text;
     private String tag;
     @ManyToOne(fetch = FetchType.EAGER)
@@ -49,11 +46,11 @@ public class Message {
         this.author = author;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
